- make timestampedbuffer use floats instead of doubles - change iso receive back to the efficient case
git-svn-id: http://subversion.ffado.org/ffado/branches/streaming-rework@397 2be59082-3212-0410-8809-b0798e1608f0

+++ src/libstreaming/IsoHandler.cpp
@@ -341,7 +341,7 @@ unsigned int IsoHandler::getCycleTimerTicks() {
     if(err) {
         debugWarning("raw1394_read_cycle_timer: %s\n", strerror(err));
     }
-    return CYCLE_TIMER_TO_TICKS(ctr.cycle_timer);
+    return CYCLE_TIMER_TO_TICKS((uint32_t)ctr.cycle_timer);
 
 #else
     // use the estimated version
